# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sulocraft API
  description: Backend service for the Sulocraft direct-to-consumer crochet storefront.
  version: 0.2.0
paths:
...
  /api/v1/payments/intent:
    post:
      tags:
      - payments
      summary: Create Payment Intent
      description: Create a gateway payment intent/order for an order.
      operationId: create_payment_intent_api_v1_payments_intent_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentIntentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/verify:
    post:
      tags:
      - payments
      summary: Verify Payment
      description: 'Verify payment completion callback from client gateway popup.


        On success, transitions Payment status to SUCCESS, Order payment_status to
        PAID,

        Order status to CONFIRMED, and records audit history.'
      operationId: verify_payment_api_v1_payments_verify_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentVerifyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/webhook/{provider}:
    post:
      tags:
      - payments
      summary: Payment Webhook
      description: Gateway webhook handler for asynchronous payment updates.
      operationId: payment_webhook_api_v1_payments_webhook__provider__post
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      - name: x-razorpay-signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Razorpay-Signature
      - name: x-signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentWebhookResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{payment_id}:
    get:
      tags:
      - payments
      summary: Get Payment
      description: Retrieve payment details.
      operationId: get_payment_api_v1_payments__payment_id__get
      parameters:
      - name: payment_id
        in: path
        required: true
        schema:
          type: integer
          title: Payment Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: x-session-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Session-Token
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaymentIntentCreate:
      properties:
        orderId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Orderid
          description: Database ID of the order
        orderNumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Ordernumber
          description: Readable order number (e.g. CB-20260930-XXXX)
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
          description: 'Payment provider: ''razorpay'' or ''mock'''
      type: object
      title: PaymentIntentCreate
      description: Request to initiate payment for an existing order.
    PaymentIntentOut:
      properties:
        paymentId:
          type: integer
          title: Paymentid
        orderId:
          type: integer
          title: Orderid
        orderNumber:
          type: string
          title: Ordernumber
        provider:
          type: string
          title: Provider
        providerOrderId:
          anyOf:
          - type: string
          - type: 'null'
          title: Providerorderid
        amount:
          type: integer
          title: Amount
        amountPaise:
          type: integer
          title: Amountpaise
        currency:
          type: string
          title: Currency
          default: INR
        keyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyid
        notes:
          additionalProperties: true
          type: object
          title: Notes
      type: object
      required:
      - paymentId
      - orderId
      - orderNumber
      - provider
      - amount
      - amountPaise
      title: PaymentIntentOut
      description: Payment intent output returned to frontend checkout.
    PaymentOut:
      properties:
        id:
          type: integer
          title: Id
        orderId:
          type: integer
          title: Orderid
        provider:
          type: string
          title: Provider
        providerPaymentId:
          anyOf:
          - type: string
          - type: 'null'
          title: Providerpaymentid
        providerOrderId:
          anyOf:
          - type: string
          - type: 'null'
          title: Providerorderid
        amount:
          type: integer
          title: Amount
        amountPaise:
          type: integer
          title: Amountpaise
        currency:
          type: string
          title: Currency
          default: INR
        status:
          type: string
          title: Status
        paymentMethodDetail:
          anyOf:
          - type: string
          - type: 'null'
          title: Paymentmethoddetail
        createdAt:
          type: string
          title: Createdat
        updatedAt:
          type: string
          title: Updatedat
      type: object
      required:
      - id
      - orderId
      - provider
      - amount
      - amountPaise
      - status
      - createdAt
      - updatedAt
      title: PaymentOut
      description: Detailed payment record representation.
    PaymentVerifyRequest:
      properties:
        paymentId:
          type: integer
          title: Paymentid
        providerPaymentId:
          type: string
          title: Providerpaymentid
          description: Gateway payment ID (e.g. pay_XXXX)
        providerOrderId:
          anyOf:
          - type: string
          - type: 'null'
          title: Providerorderid
          description: Gateway order ID
        providerSignature:
          anyOf:
          - type: string
          - type: 'null'
          title: Providersignature
          description: HMAC verification signature
        paymentMethodDetail:
          anyOf:
          - type: string
          - type: 'null'
          title: Paymentmethoddetail
          description: e.g. UPI / Google Pay
      type: object
      required:
      - paymentId
      - providerPaymentId
      title: PaymentVerifyRequest
      description: Client verification callback payload after gateway popup completion.
    PaymentWebhookResult:
      properties:
        status:
          type: string
          title: Status
          default: ok
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: PaymentWebhookResult
      description: Webhook acknowledgment response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...